Project article on ACP id from AA seq
The structure‑based cancer‑related single amino acid variation prediction (2021)
. In this investigation, all SAV characteristics generated from protein sequences,
structures and the microenvironment were converted into feature vectors and fed into an integrated
predicting system using a support vector machine and genetic algorithm. Critical features were used
to estimate the relationship between their properties and cancers caused by SAVs. We describe
how we developed a prediction system based on protein sequences and structure that is capable of
distinguishing if the SAV is related to cancer or not. The fve-fold cross-validation performance of our
system is 89.73% for the accuracy, 0.74 for the Matthews correlation coefcient, and 0.81 for the F1
score. We have built an online prediction server, CanSavPre (http://bioinfo.cmu.edu.tw/CanSavPre/),
which is expected to become a useful, practical tool for cancer research and precision medicine.