Id subtypes of ACPs based on sequential features
Identifcation of subtypes of anticancer peptides based on sequential features (2021)
. In this study, we compared the N- and C-terminal
amino acid composition for each peptide, there are three major subtypes of ACPs that are defned
based on the distribution of positively charged residues. For the frst time, we were motivated to
develop a two-step machine learning model for identifcation of the subtypes of ACPs, which classify
the input data into the corresponding group before applying the classifer. Further, to improve
the predictive power, the hybrid feature sets were considered for prediction. Evaluation by fvefold cross-validation showed that the two-step model trained with sequence-based features and
physicochemical properties was most efective in discriminating between ACPs and non-ACPs. The
two-step model trained with the hybrid features performed well, with a sensitivity of 86.75%, a
specifcity of 85.75%, an accuracy of 86.08%, and a Matthews Correlation Coefcient value of 0.703.
Furthermore, the model also consistently provides the efective performance in independent testing
set, with sensitivity of 77.6%, specifcity of 94.74%, accuracy of 88.99% and the MCC value reached
0.75. Finally, the two-step model has been implemented as a web-based tool, namely iDACP, which is
now freely available at http://mer.hc.mmh.org.tw/iDACP/